added custom material binder and fixed bug

+++ sht/graphics/src/image/image_hdr.cpp
@@ -17,7 +17,7 @@ namespace {
 	float ConvertComponent(int expo, int val)
 	{
 		float v = val / 256.0f;
-		float d = powf(2.0f, expo);
+		float d = powf(2.0f, (float)expo);
 		return v * d;
 	}
